backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
bcrypt==4.0.1
python-dotenv==1.0.0
werkzeug==2.3.7
PyPDF2==3.0.1
Pillow==10.0.1
user-agents==2.2.0
qrcode==7.4.2
redis==5.0.1
celery==5.3.4
flask-limiter==3.5.0
python-magic==0.4.27
